Add a common file for utils
