+++ packages/dd-trace/src/appsec/telemetry/common.js
@@ -1,0 +1,24 @@
+'use strinct'
+
+const tags = {
+  BLOCK_FAILURE: 'block_failure',
+  EVENT_RULES_VERSION: 'event_rules_version',
+  INPUT_TRUNCATED: 'input_truncated',
+  REQUEST_BLOCKED: 'request_blocked',
+  RULE_TRIGGERED: 'rule_triggered',
+  WAF_ERROR: 'waf_error',
+  WAF_TIMEOUT: 'waf_timeout',
+  WAF_VERSION: 'waf_version'
+}
+
+function getVersionsTags (wafVersion, rulesVersion) {
+  return {
+    [tags.WAF_VERSION]: wafVersion,
+    [tags.EVENT_RULES_VERSION]: rulesVersion
+  }
+}
+
+module.exports = {
+  tags,
+  getVersionsTags
+}
